Split survey actions file up
